+++ src/arkreactor/Compiler/Node.cpp
@@ -202,6 +202,7 @@ namespace Ark::internal
                     case Keyword::Import: os << "Import"; break;
                     case Keyword::Quote: os << "Quote"; break;
                     case Keyword::Del: os << "Del"; break;
+                    case Keyword::Pop: os << "Pop"; break;
                 }
                 break;