+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased changes]
 ### Added
+- internal hidden keyword `Pop` to use in the stack trashing
 
 ### Changed
 - using `doc_formatting.first_column` instead of `doc_formatting.start_column` when displaying the CLI help